fix: impersonate logout
We have a hackish way to adjust logout for impersonate users which
relies on the HTML structure. On major releases this changed, so the
original logout logic was in place.

The selectors were adjusted, and also wait until the target elements are
actually available. Use of jQuery was removed along.

Signed-off-by: Arthur Schiwon <blizzz@arthur-schiwon.de>